lease-saas-all-in-one-final/server/requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
sqlalchemy
psycopg[binary]
python-multipart
passlib[bcrypt]
pyjwt
pypdf
langextract
stripe
redis
prometheus-client
prometheus-fastapi-instrumentator
